refactor: simplify i18n-related code
