+++ src/configI18n.js
@@ -1,34 +1,10 @@
-import { i18nKeys } from './i18n'
-import i18n from './locales'
-
-const isLangRTL = code => {
-    const langs = ['ar', 'fa', 'ur']
-    const prefixed = langs.map(c => `${c}-`)
-    return (
-        langs.includes(code) ||
-        prefixed.filter(c => code.startsWith(c)).length > 0
-    )
-}
-
-export const configI18n = userSettings => {
-    const lang = userSettings.keyUiLocale
-    if (isLangRTL(lang)) {
-        document.body.setAttribute('dir', 'rtl')
-    }
-
-    i18n.changeLanguage(lang)
-}
+import i18n from '@dhis2/d2-i18n'
+import { i18nKeys } from './i18n-keys'
 
 export const injectTranslationsToD2 = d2 => {
-    if (d2) {
-        const translations = {}
-        const translationKeys = Object.keys(i18nKeys.d2UiComponents)
-        for (let i = 0; i < translationKeys.length; i++) {
-            const key = translationKeys[i]
-            translations[key] = i18n.t(i18nKeys.d2UiComponents[key])
-        }
-        Object.assign(d2.i18n.translations, translations)
-    }
+    const translations = {}
+    Object.keys(i18nKeys.d2UiComponents).forEach(key => {
+        translations[key] = i18n.t(i18nKeys.d2UiComponents[key])
+    })
+    Object.assign(d2.i18n.translations, translations)
 }
-
-export default configI18n
