+++ src/pages/data-integrity/data-integrity-card/DataIntegrityCard.js
@@ -1,8 +1,8 @@
+import i18n from '@dhis2/d2-i18n'
 import { FontIcon } from 'material-ui'
 import { Card, CardText, CardHeader } from 'material-ui/Card'
 import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
-import i18n from '../../../locales'
 import styles from './DataIntegrityCard.module.css'
 
 const jsStyles = {
